requirements.txt
# Core deep learning
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Ultralytics YOLO (pose, detection, etc.)
ultralytics>=8.3.78

# Experiment tracking
mlflow>=2.12.0
dagshub>=0.3.13

# Data science stack
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
python-box==6.0.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Video & image processing
opencv-python>=4.9.0.80

# Optional for notebooks
jupyter>=1.0.0
ipykernel>=6.29.0

# Local project (editable install)
-e .